pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "python-dotenv",
    "google-api-python-client",
    "google-generativeai",
    "tenacity",
    "isodate",
    "watchfiles",
]